Split config into project and local options
