+++ gulp-config.json.sample
@@ -1,6 +1,5 @@
 {
 	"wwwDir"           : "/var/www/jcms3x",
 	"browserSyncProxy" : "localhost",
-	"extensions"       : {
-	}
+	"browser"          : "default"
 }
